requirements.txt
Flask>=3.0.3
tensorflow>=2.16.0
Pillow>=10.0.0
numpy>=1.24.0
gunicorn>=23.0.0